It's the eve of 2020, almost 7 years since 0x10c was a thing.
However I still remember just how much excitement and interest this thing stired up in not just retro
hungry developers. But also myself personally. The DCPU-16 came around right as I was starting to seriously
become both interested and capable of understanding the inner workings and implementation of an emulator,
and the vast amount of community interest provided a wealth of modern resources and programs that targeted
an architecture.

Even currently the source code of assemblers and compilers targetting this platform is a place where I start when
looking to explore how I would implement those things.

Unfortunately 0x10c died, and the community around it's hardware quickly died with it. But instead of abandoning the
software, I think it should be a place to learn. RISC-V is arguably a better (and more pratical) place to start for
working on a project like this these days. The DCPU-16 architecture is quirky in a somewhat pointless way.
16-bit cpus are actually slightly harder to emulate on modern 64-bit machines in some ways.

A RISC-V assembler, emulator or compiler project would result in both a far more pratical knowledge
and software than some 7 year old architecture for a dead game.

But still the fact that there's only like 3-4 standard peripherals, with simple, emulation focused interfaces, and a
very limited and well defined system.